move algolia to registry_search.ts
