+++ pages/x/index.tsx
@@ -10,7 +10,8 @@ import TimeAgo from "timeago-react";
 import Header from "../../components/Header";
 import Footer from "../../components/Footer";
 import InlineCode from "../../components/InlineCode";
-import { listModules, getStats } from "../../util/registry_utils";
+import { getStats } from "../../util/registry_utils";
+import { listModules } from "../../util/registry_search";
 import * as pageutils from "../../util/pagination_utils";
 import RegistryInstructions from "../../components/RegistryInstructions";
 import { CookieBanner } from "../../components/CookieBanner";
